ImGui asset browser, filesystem cleanup and creation of ImGui images
The filesystem has gotten an overhaul so it is more robust and has a way to handle an entire file directory with subfolders.

Added a helper for directory panels.

And I also got an AssetBrowser that is up and working now on the imported assets. Now is time to create the actual import pipeline.