Shader reflection for shaders.
Materials now serialize a shader.
Shaders are a resource that can be imported.
Shaders get reflected which makes the material editor more dynamic.
Shaders have the beginning of a hot-reloading feature that should be added to textures.
Moved around a lot of systems and made some design decisions regarding the workings of material data.