Font rendering
Mesh shader support in Vulkan and material creation.
Some Growing array functionality.
TTF font parsing
Taking in an outside library for triangulation of glyph data.
Material rework so that materials have a pipeline, so separating out the rendering responsibility of the material to a pipeline obnject.
Adding a context state object for rendering handling. This will be worked in more to the rendering in coming updates.