Reworked textures to be managed by VMA and framebuffer rework
The changes have moved (almost) all memory management to be under the context, and how I manage the vulkan memory so it is more centralized. This simplifies a lot of the code, but it is not done yet, I still need to move Imgui to the new way.

I also added a new resource state enum to make loading of all materials and textures, they are now staggered, aka loads get queued, so this will make threaded loading easier in the future,