First pass at rendering clouds
Moved fullscreen vertex buffer to be owned by the context, and it has a dedicated draw call to create a fullscreen pass.
Fixed some bugs with inconsistent Render target resolution and the order that is set up in.
Added support for Samplers as separate objects and when rendering to specify the sampler to use when sampling textures.
Added a cloud rendering class which for now hard codes a point set, which we later render in the render loop.